Create a separate place for WP hooks and move existing hook to it
