+++ packages/js/product-editor/src/wp-hooks/index.ts
@@ -1,0 +1,8 @@
+/**
+ * Internal dependencies
+ */
+import registerHideInventoryAdvancedCollapsible from './hide-inventory-advanced-collapsible';
+
+export default function registerProductEditorHooks() {
+	registerHideInventoryAdvancedCollapsible();
+}
